python-server/requirements.txt
Flask-SocketIO
pystache